Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,9 +1,9 @@
 3include<stdio.h>
 int main()
 {
-    float point;
-    scanf("%le",&point);
-    printf("You entered: %le",point)
+    double point;
+    scanf("%f",&point);
+    printf("You entered: %f",point)
     return 0;
-    
+
 }
